# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ErrorResp:
    properties:
      message:
        type: string
    type: object
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        type: string
      stack:
        description: Hide in production
    type: object
...
  dto.ApiResponse-array_dto_DiscountListItem:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DiscountListItem'
        type: array
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.ApiResponse-array_dto_DiscountRuleDetail:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DiscountRuleDetail'
        type: array
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-dto_DiscountDetail:
    properties:
      data:
        $ref: '#/definitions/dto.DiscountDetail'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  dto.ApiResponse-dto_DiscountRuleDetail:
    properties:
      data:
        $ref: '#/definitions/dto.DiscountRuleDetail'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-string:
    properties:
      data:
        type: string
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.DiscountDetail:
    properties:
      code:
        type: string
      createdAt:
        type: string
      description:
        type: string
      discountType:
        type: string
      discountValue:
        type: number
      id:
        type: string
      isActive:
        type: boolean
      maxDiscount:
        type: number
      minPurchase:
        type: number
      timesUsed:
        type: integer
      updatedAt:
        type: string
      usageHistory:
        items:
          $ref: '#/definitions/dto.DiscountUsageHistory'
        type: array
      usageLimit:
        type: integer
      validFrom:
        type: string
      validUntil:
        type: string
    type: object
  dto.DiscountListItem:
    properties:
      code:
        type: string
      createdAt:
        type: string
      description:
        type: string
      discountType:
        type: string
      discountValue:
        type: number
      id:
        type: string
      isActive:
        type: boolean
      maxDiscount:
        type: number
      minPurchase:
        type: number
      updatedAt:
        type: string
      usageLimit:
        type: integer
      usedCount:
        type: integer
      validFrom:
        type: string
      validUntil:
        type: string
    type: object
  dto.DiscountRuleDetail:
    properties:
      id:
        type: string
      ruleType:
        type: string
      ruleValue:
        additionalProperties: true
        type: object
    type: object
  dto.DiscountUsageHistory:
    properties:
      amount:
        type: number
      customerName:
        type: string
      date:
        format: "2006-01-02"
        type: string
      discountAmount:
        type: number
      id:
        type: string
      orderId:
        type: string
    type: object
  dto.ErrorResp:
    properties:
      error:
        $ref: '#/definitions/dto.ApiError'
    type: object
...
  dto.MetaInfo:
    properties:
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  models.AddDiscountModel:
    properties:
      code:
        maxLength: 32
        minLength: 5
        type: string
      description:
        maxLength: 1000
        type: string
      discountType:
        type: string
      discountValue:
        type: number
      isActive:
        type: boolean
      isStackable:
        type: boolean
      maxDiscountAmount:
        type: number
      minOrderValue:
        type: number
      name:
        maxLength: 100
        minLength: 3
        type: string
      priority:
        minimum: 0
        type: integer
      usageLimit:
        minimum: 0
        type: integer
      validFrom:
        type: string
      validUntil:
        type: string
    required:
    - code
    - discountType
    - discountValue
    - isActive
    - name
    - validFrom
    type: object
  models.AddDiscountRuleModel:
    properties:
      ruleType:
        enum:
        - product
        - category
        - customer_segment
        - brand
        - first_time_buyer
        - purchase_quantity
        type: string
      ruleValue:
        additionalProperties: true
        type: object
    required:
    - ruleType
    - ruleValue
    type: object
...
  models.UpdateDiscountModel:
    properties:
      code:
        maxLength: 32
        minLength: 5
        type: string
      description:
        maxLength: 1000
        type: string
      discountType:
        type: string
      discountValue:
        type: number
      isActive:
        type: boolean
      isStackable:
        type: boolean
      maxDiscountAmount:
        type: number
      minOrderValue:
        type: number
      name:
        maxLength: 100
        minLength: 3
        type: string
      priority:
        minimum: 0
        type: integer
      usageLimit:
        minimum: 0
        type: integer
      validFrom:
        type: string
      validUntil:
        type: string
    type: object
  models.UpdateDiscountRuleModel:
    properties:
      ruleType:
        enum:
        - condition
        - action
        type: string
      ruleValue:
        additionalProperties: true
        type: object
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: This is a sample server for a simple e-commerce API.
  title: E-Commerce API
paths:
...
  /admin/discounts:
    post:
      consumes:
      - application/json
      description: Create a new discount
      parameters:
      - description: Discount info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddDiscountModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_DiscountDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Create a new discount
      tags:
      - discounts
  /admin/discounts/{id}/rules:
    get:
      consumes:
      - application/json
      description: Get all discount rules for a specific discount
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-array_dto_DiscountRuleDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Get all discount rules for a discount
      tags:
      - discounts
    post:
      consumes:
      - application/json
      description: Add a discount rule to a discount
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      - description: Discount rule info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.AddDiscountRuleModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ApiResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Add a discount rule to a discount
      tags:
      - discounts
  /admin/discounts/{id}/rules/{ruleId}:
    delete:
      consumes:
      - application/json
      description: Delete a discount rule
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Delete a discount rule
      tags:
      - discounts
    get:
      consumes:
      - application/json
      description: Get a specific discount rule by ID
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_DiscountRuleDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Get a specific discount rule by ID
      tags:
      - discounts
    put:
      consumes:
      - application/json
      description: Update a discount rule
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      - description: Rule ID
        in: path
        name: ruleId
        required: true
        type: string
      - description: Updated discount rule info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDiscountRuleModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Update a discount rule
      tags:
      - discounts
...
  /discounts:
    get:
      consumes:
      - application/json
      description: Get all discounts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Search by code
        in: query
        name: search
        type: string
      - default: percentage
        description: Discount type
        in: query
        name: discountType
        type: string
      - default: true
        description: Is active
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-array_dto_DiscountListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Get all discounts
      tags:
      - discounts
  /discounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete discount by ID
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Delete discount by ID
      tags:
      - discounts
    get:
      consumes:
      - application/json
      description: Get discount by ID
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_DiscountDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Get discount by ID
      tags:
      - discounts
    put:
      consumes:
      - application/json
      description: Update discount by ID
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      - description: Discount info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDiscountModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Update discount by ID
      tags:
      - discounts
...
swagger: "2.0"